score card ui design complete
