+++ lib/main.dart
@@ -13,6 +13,4 @@ class MyApp extends StatelessWidget {
       home: HomePage(),
     );
   }
-
-
 }
